Use Rc for arrays to improve performance
Every time we access a variable, the Context::get_var function is
called. This function clones the value of the variable.

When the variable is an array, this is an expensive operation since
every element in the array needs to be cloned. By wrapping the
Vec<RickrollObject> of an array in an Rc, we now only need to clone
a reference to the vector when accessing the array, which makes the
operation much faster.